requirements.txt
# KYC Multi-Provider Document Processor
# Install: pip install -r requirements.txt

click>=8.1,<9
openai>=1.30,<2
pydantic>=2.5,<3
python-dotenv>=1.0,<2
Pillow>=10.0,<12
PyYAML>=6.0,<7

# Optional: native Google Gemini SDK (falls back to OpenAI-compat without it)
# google-genai>=1.0

# Development / testing
pytest>=8.0,<9
pytest-cov>=5.0,<6
ruff>=0.4,<1
